+++ frontend/app_settings.py
@@ -21,12 +21,17 @@ from utils.schema import AppInfra
 PAGE_ICON = "./datarobot_favicon.png"
 
 
-def get_page_logo() -> str:
+def display_page_logo() -> None:
     theme = st_theme()
-    logo = "./DataRobot_white.svg"
-    if theme and theme.get("base") == "light":
-        logo = "./DataRobot_black.svg"
-    return logo
+    # logo placeholder used for initial load
+    logo = '<svg width="133" height="20" xmlns="http://www.w3.org/2000/svg" id="datarobot-logo"></svg>'
+    if theme:
+        if theme.get("base") == "light":
+            logo = "./DataRobot_black.svg"
+        else:
+            logo = "./DataRobot_white.svg"
+    with st.container(key="datarobot-logo"):
+        st.image(logo, width=200)
 
 
 def get_database_logo(app_infra: AppInfra) -> None: