+++ infra/settings_database.py
@@ -12,7 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
 
 from utils.schema import DatabaseConnectionType
 
-DATABASE_CONNECTION_TYPE: DatabaseConnectionType = "snowflake"
+# Valid values are: "snowflake", "bigquery", or "no_database"
+DATABASE_CONNECTION_TYPE: DatabaseConnectionType = os.getenv(  # type: ignore
+    "DATABASE_CONNECTION_TYPE", "no_database"
+)