+++ tests/e2e/test_run_app.py
@@ -23,15 +23,15 @@ from tests.e2e.utils import (
     wait_for_element_to_be_visible,
 )
 
-PROCESSING_TIMEOUT = 120
+PROCESSING_TIMEOUT = 240
 
 
 def assert_data_processed(browser) -> None:
     assert wait_for_element_to_be_visible(
         browser,
         By.XPATH,
-        "//p[contains(text(), '✅ Data processed and dictionaries generated successfully!')]",
-        120,
+        "//p[contains(text(), 'Data processed and dictionaries generated successfully!')]",
+        PROCESSING_TIMEOUT,
     )
 
 
@@ -122,6 +122,30 @@ def test_data_dictionary_loaded(browser: webdriver.Chrome, get_app_url: str) -> 
 
     assert wait_for_element_to_be_visible(
         browser, By.XPATH, "//p[contains(text(), 'Download Data Dictionary')]"
+    )
+
+
+@pytest.mark.usefixtures("check_if_logged_in")
+def test_clear_data_button(browser: webdriver.Chrome, get_app_url: str) -> None:
+    browser.get(get_app_url)
+    load_from_database(browser, "LENDING_CLUB_PROFILE")
+
+    click_element(
+        browser,
+        By.XPATH,
+        "//p[contains(text(), 'Clear Data')]",
+    )
+
+    click_element(
+        browser,
+        By.XPATH,
+        "//span[contains(text(), 'Data Dictionary')]",
+    )
+
+    assert wait_for_element_to_be_visible(
+        browser,
+        By.XPATH,
+        "//p[contains(text(), 'Please upload and process data from the main page to view the data dictionary')]",
     )