+++ CHANGELOG.md
@@ -7,7 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [0.1.3] - 2025-02-03
+## [0.1.5] - 2025-02-12
+
+## Added
+- LLM tool use support
+- Checkboxes allow changing conversation
+- DATABASE_CONNECTION_TYPE can be set from environment
+
+### Fixed
+- Fix issue where plotly charts reuse the same key
+- Fix [Clear Data] button
+- Fix logo rendering on first load
+- Fix Data Dictionary editing
+
+## [0.1.4] - 2025-02-03
 
 ### Changed
 - Better cleansing report, showing more information
@@ -41,3 +54,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial functioning version of Pulumi template for data analyst
 - Changelog file to keep track of changes in the project.
 - pytest for api functions
+